+++ src/server/resolvers/VolunteerResolver.ts
@@ -1,0 +1,8 @@
+import Context from '../context';
+import { VolunteerResolvers, UserType } from '../generated/graphql';
+import { Hacker } from './HackerResolver';
+
+export const Volunteer: VolunteerResolvers<Context> = {
+	...Hacker,
+	userType: () => UserType.Volunteer,
+};
